scripts/Flask-Docker-App/requirements.txt
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
Werkzeug==2.0.2
scikit-learn==1.2.2
numpy==1.24.2
sklearn-crfsuite==0.3.6
pyvi==0.1.1
